GS-9054
Fix class cast in BootstrapUtility

svn path=/trunk/openspaces/; revision=91911


Former-commit-id: 21ac913528d0a7a3694fd0f41840a5d0eda6d9c8

+++ src/main/src/org/openspaces/core/gateway/GatewaySinkFactoryBean.java
@@ -171,12 +171,12 @@ public class GatewaySinkFactoryBean extends AbstractGatewayComponentFactoryBean 
         if (namedArgs.containsKey("bootstrapFromGateway"))
         {
             String bootstrapRemoteGatewayName = (String) namedArgs.get("bootstrapFromGateway");
-            Integer bootstrapTimeoutInSeconds = (Integer) namedArgs.get("bootstrapTimeout");
+            Long bootstrapTimeoutInSeconds = (Long) namedArgs.get("bootstrapTimeout");
             try 
             {
                 BootstrapConfig config = new BootstrapConfig(bootstrapRemoteGatewayName);
                 if (bootstrapTimeoutInSeconds != null)
-                    config.setTimeout(bootstrapTimeoutInSeconds.intValue());
+                    config.setTimeout(bootstrapTimeoutInSeconds.longValue());
                 
                 localClusterReplicationSink.bootstrapFromRemoteSink(config);
             } catch (Exception e) {
